Fix PTL-Approved column in RelMgt dashboard
The PTL-Approved label is currently set as a submit requirement, so that
it appears in the review dashboards. However, in order to not block the
submit, that requirement is always satisfied (the default of 0 is above
MIN), so that column is always shown with a green "Satisfied" checkmark.

This change makes PTL-Approved a blocking submit requirement (the
default of 0 now being the MIN) while allowing the release managers to
set +1 in extreme cases where PTL approval cannot be obtained. This
should ensure that the green checkmark in the PTL-Approved column works
correctly and lets us spot which changes can be processed quickly.

Change-Id: I9d99ba7edb996e871e8bc8d2cc1c3db0197a4b9e