Prevent recreate EOL'd branch
In case a branch is End of Life (marked with either $series-eol or
bugfix-<version>-eol) and was deleted before, then our script recreates
it based on the info in deliverable's yaml file.

This patch extends the script by checking if an *-eol tag exists and
in case it does then exits without creating again the already deleted
branch.

Change-Id: I1b763aa52703221d211156c060b8998282fcfc5d