Remove retied projects from openstack-helm group
At the moment we have 3 active projects in openstack-helm
project group. These are
- openstack-helm
- openstack-helm-infra
- openstack-helm-images
But we used to have 3 more projects in the project group
which are now retired. These retied projects are
- openstack-helm-docs
- openstack-helm-addons
- openstack-helm-deployments
Now we would like to cleanup the openstack-helm
Storyboard project group. We have closed and moved all the
stories from these retired projects to the main OSH project.

Change-Id: Ib60099f79de902906b8d1956cc202aeb86587cb5