Update the gerrit acl for retired winstacker project
Winstackers project has been retired
- https://review.opendev.org/c/openstack/governance/+/886880

There is no maintainer left for this project so
we need to update the acls to remove the content of
all the deliverables of this project.

Rest of the infra cleanup of this projects will
be done in separate commit after the repo content
reomval.

Change-Id: I18b59d3b72897bac67967c1c2a9d3c0412a56b49