Remove ara from Zuul config
As far as I can tell this was only used by opendev/system-config to do
integration testing between opendev playbooks and ara development. We
don't really need this as the important piece of that job is checking
development ansible works. There is some motivation to reduce the number
of gitub repos in zuul without apps installed to avoid github api rate
limits when starting zuul from a clean slate. Ara is an easy first
cleanup so we start here.

Depends-On: https://review.opendev.org/c/opendev/system-config/+/894576
Change-Id: I58120c48b6db2dc5c8d4ecb7e5791d42cb35ef2f