Remove two unnused github projects from tenant config
We discovered we hit github api rate limits trying to load projects
without the github zuul app installed. That led to me doing some
investigation on whether or not these repos are needed at all.

Best I can tell the zookeeper-operator repo was looked at for us with
the zuul operator and is not longer needed. Eventlet was added for use
by swift [0], but in the related change [1] they only used the project
in older patchsets. Newer patchsets (including the one that merged) did
not use the required-project at all making this unneeded.

[0] https://review.opendev.org/c/openstack/project-config/+/656889
[1] https://review.opendev.org/c/openstack/swift/+/653548

Change-Id: I8e5cab01850969f29c78a165686502b948f2d621