Add benchmark subteam to charms teams
We have a distinct set of contributors who we'd like to engage in
the subset of charms that support benchmark testing rather than
direct deployment of OpenStack.

Create a new set of 'charms-benchmark' team ACL's to support this
and add to charm-magpie.

A second charm (woodpecker) is currently proposed for inclusion
into the project and will sit under this ACL as well.

Change-Id: Iff4811964a28436a77a3e487d966ccc10a5fab47