Temporarily remove release docs semaphores
During OpenStack coordinated releases, hundreds of release
documentation and release note publication jobs are triggered, but
with the semaphores end up taking most of a day because they all get
serialized. These semaphores are intended to avoid builds for
different branches of the same project from racing one another and
publishing out of sequence, but on release day only tags for a
single branch (master) are triggered making the semaphores
unnecessary in that situation.

We temporarily take out the semaphores prior to starting the
coordinated release process for OpenStack, and then revert this
change once the jobs have completed.

Change-Id: Ie5a3d81a3d99acb3c11279747e94405546fd6d74