Check proposed project names for problems
Today we discovered that Gitea doesn't allow creating projects with
"+" in their names (or that it needs special escaping in some way).
In order to avoid future incidents, let's be defensive and limit
organization/project names to only the characters which are
currently in use, since we know those work. If people want to use
characters outside that list in the future, we can test for
potential breakage before amending the filter.

Change-Id: I8b87ab6d52e0ec7da92f80a62c7d97e4f2245b29