Update secret for Weblate REST API key
I have added an infra-root E-mail alias for a new
weblate@openstack.org account in id.openinfra.dev, created an
account in openstack.weblate.cloud with that ID, and used it to
generate a Weblate REST API key. This change updates the previous
placeholder secret with one containing the new key.

Change-Id: I88c5ee4d3847fc4a59130f746cbfb8609df08939