Fix wheel_volume values for centos stream wheel mirrors
We were trying to vos release the non stream centos 8 volumes which
don't exist anymore. Flip this to the stream values. Additionally,
centos 9 stream arm64 was releasing x86_64 wheels not arm64 wheels. Fix
this too.

This probably is an indication that these mirrors were never really used
much in the first place and we should consider cleanup?

Change-Id: I7f4104a342d170669e7717b945f09131811dbebc