Add warning to nodepool configs about changing cloud name
The cloud name is used to lookup cloud credentials in clouds.yaml, but
it is also used to determine names for things like mirrors within jobs.
As a result changing this value can impact running jobs as you need to
update DNS for mirrors (and possibly launch new mirrors) first. Add a
warning to help avoid problems like this in the future.

Change-Id: I9854ad47553370e6cc9ede843be3303dfa1f9f34