Remove last bit of system-config-puppet-apply-jobs usage
This was missed when doing an earlier cleanup. This is part of the
Ubuntu Xenial removal. We have to remove puppet apply jobs for
system-config because they all rely on puppet on Xenial. We don't retire
these projecst and instead switch them to noop jobs because we may still
need to make changes to them (though we hope we don't).

Change-Id: Iba1b4ffbeeecfc39b6bd17db829f7372f605b8b2