Add devstack-gate back to central config for retirement prep
Several years ago all of devstack-gate's zuul config was moved into
devstack-gate proper. We need to keep the official jobs that do github
syncing here in project-config and we need to add noop-jobs to
project-config in order to "remove" the repo content in preparation for
its retirement. Do that here.

Followup changes will retire the repo then eventually make it back to
this repo to remove the config I have just added.

Change-Id: Ib7aa64663cd6306e69b49e15e80419230079a8a8