Add editHashtags permissions to some governance repositories
The TC uses topic tags to identify kinds of changes and applies
rules based on the topic; so if contributors want to add further
filtering metadata that's useful, they can use hashtags instead.
Currently, allow the respective core group to use hashtags. This
may just change to Registered Users in the future if the infra team
finds that more appropriate.

Change-Id: I02933d6a72e295f237ada1bc012eff99d7b25573