Add wandertracks repos
WanderTracks is an Open Source project for collecting and tracking
information journeys taken.

The intent is for it to support both normal (walking, hiking, driving)
and specialty (sailing, diving) activities, along with collecting
technical information about vessels involved where appropriate. The
existing systems are either hyper-focused on one veritcal, like
saillogger for sailing, or TelsaMate for driving a Tesla, or are *super*
generic GPS trackers, like Traccar, which are great but are essentially
key-value stores.

WanderTracks will allow importing and managing trips either directly
from live collectors, or via import from geojson files.  The web
interface will support displaying individual trips or collections of
trips overlaid onto a map. The very-very manual proof of concept of the
collection of trips display can be seen at
https://waterwanders.com/where ... wandertracks is going to be the
generalized and multi-user version of that.

To keep the work on infra-root to a minimum, create all the repos we're
going to need up front:

 wandertracks/system-config - Ansible for Zuul CD
 wandertracks/signalk-wandertracks - SignalK Data collection Plugin
 wandertracks/wandertracks - API service
 wandertracks/wandertracks-web - Web UI
 wandertracks/wanders.cloud - Multi-tenant site built on wandertracks
 waterwanders/waterwanders - Single-tenant site using backend APIs

Change-Id: I72f0491fa96c97cfac1bf20e6b92912dd087b589