Set inmotion max servers to 0
This is the first step in winding down this cloud. We plan to redeploy
the cloud on top of an entirely new platform and openstack version. We
can't do that in a new cloud adjacent to the existing one; instead we
must tear down the existing cloud and redeploy on the same hardware.
Wind it down gracefully in Nodepool first.

Change-Id: Icdaca06b2a2737f74daf60cb29674f06160c7faf