Remove the inmotion cloud entirely
This is the last step in inmotion cloud cleanup. It does leave nl02 as a
nodepool launcher with no active providers. I suspect this is fine and
we'll add the new OpenMetal cloud to nl02 at some point in the future.

The grafana graphs will also need to be manually deleted at some point
as removing the yaml file doesn't remove the dashboard from grafana.

Change-Id: Ib33e0c45c277f77013fe5820b898df03da58b558