Add OpenMetal to Nodepool and Grafana
This is essentially reverting commits bd15ddc and cb4b99b which were
the final stages of winding down and cleaning up the old InMotion
cloud which OpenMetal has replaced, with the cloud name updated (but
region kept the same) and grafyaml data regenerated. It stops short
of actually booting nodes in the new environment until we have a
chance to spot check things once images get uploaded.

Since this is re-adding diskimages back to nl02, I refrained from
including centos-8-stream which is in the progress of being removed,
so that we don't unnecessarily upload images we're not planning to
boot.

Change-Id: If8e9b7105b4c7a13e87ebb4f6c985e821c30a842