Puppet OpenStack: Drop project group
Each individual repo of Puppet OpenStack project has its own launchpad
project. Drop the project group so that each project is linked to
the appropriate project.

Note that the group override is kept for puppet-openstack_extras but
adjusted, because of difference between the repository name (_extras)
and the project name (-extras).

Change-Id: Iabaaaf678ccb1c2e08999268fc0f4937bb93437e