Fix image uploads for openmetal
The images for centos-8-stream and gentoo are no longer buildable and
nodepool deleted the old "raw" versions of it in order to save space, so
these currently cannot be uploaded to the new openmetal cloud.

Add a reduced list of images as a temporary workaround, this can be
reverted once the old images have been cleaned out globally.

Also add an explicit "pause: true" stanza to the gentoo image build
which so far has only been paused manually.

Change-Id: I4c67940cb0ff9ead6ae2274e00e6f48e7cbcce89