nodepool: revert dedicated diskimages list for rax-iad
When we had some issues with image uploads to rax-iad last year, we
needed to configure it with a dedicated list of diskimages in order to
be able to add "pause: true" statements to those. When the issue was
resolved, we only dropped the pause statements instead of reverting to
use the original list shared across all providers [0]. This is getting
fixed now.

[0] Ia2b737e197483f9080b719bab0ca23461850e157

Change-Id: Ibc1a11d04d5258ca503f8cd5b5fe7a30eb99bc56