nodepool: reorder providers to list rax-dfw last
There are currently some image upload issues for the rax-dfw region.
Since there is an issue in nodepool that causes uploads to other
provider getting blocked by this (see [0]), move the rax providers to
the end of the list, so all other providers will be updated before
hittinh the issue.

[0] https://review.opendev.org/c/zuul/nodepool/+/922242

Change-Id: Ib8bec7ea2afd5b4569bddb9a4feea6bb07a280f3