Limit what volvocars loads from opendev/base-jobs
On changes like https://review.opendev.org/c/opendev/base-jobs/+/922653
the volvocars tenant is running jobs for the base-jobs repo because that
tenant is loading the project definition for base-jobs. We really only
want the opendev tenant to be running those jobs so we limit what we
include in the volvocars tenant to jobs, secrets, and nodesets. This
matches what we've done in other tenants like zuul and pyca.

While we are in there we update the comment associated with this config
in other tenants to make it a bit more accurate.

Change-Id: Iea24607def728bfd170a6f43da2eb4960d0c89e4