Set linaro cloud's max servers to 0
The hardware that the linaro cloud is running on is going away on
August 10. In preparation for that shutdown we're going to wind down our
use of the cloud now. The first step in doing that is to set the maximum
test node usage to 0. Once nodes are no longer used we can clean up
images and so on.

Change-Id: Ib4655908190869b255807bd9a42645be4c339fb0