Remove labels and diskimages from the linaro cloud
This is the next step in cleaning up the linaro cloud. We set the
diskimages and labels lists to empty list in the provider config for
this cloud. This will have the affect of cleaning up the diskimages in
that cloud region from both the cloud and the related records in the zk
database. This should be landed after the max-servers: 0 change and
we've manaully deleted any existing nodes.

The next change will remove the provider config entirely.

Change-Id: I8ee50779637e980132cf5163e4b18577bcfb3bec