Use the osuosl mirror for deb packages in image builds
The linaro cloud will be shutdown in the near future which means its
mirror will go away. Stop using its mirror for deb package content and
pull those packages from the osuosl mirror instead. This has the added
benefit of fetching the packages from the mirror that is local to the
builder which should be more efficient.

Change-Id: I5ea798faa0cb571834050f0082c7cf782e2641df