Remove linaro cloud from Nodepool
This should be done after we are happy with the cleanups that nodepool
has performed particularly against the zk database. Otherwise we may
need to do manual cleanup after this provider config is removed.

We also remove the auto generation of the linaro cloud grafana dashboard
rules but don't remove the dashboard itself. We typically keep the
dashboards around for a while as the information exposed there may
continue to be useful after the cloud is gone.

Change-Id: I40acd28146484cb81f87ee825787d578eb964de8