Update update-constraints job and script
We often see failures on requirements patches generates after a release
was tagged that are due to the newly created version not being found yet
in the index. These require a recheck to resolve. Add a pause of 10
minutes before creating the requirements review in an attempt to reduce
the frequency of these failures.

Also fix some minor issues:

- The extraction of the SHORTNAME for the repo from remote.origin.url
  git config is broken, because zuul overrides that url to be /dev/null.
  Use the working directory name instead.
- Current python minor versions have two digits, amend the regex we are
  using to match these.
- Drop a workaround that was needed when Python 3.8 was new and shiny.
- Switch the nodeset this job is running on from ubuntu-focal to
  ubuntu-noble.

Change-Id: I6751ea91499bbc81c71dce6da7fbe982bbc03efe