Use yaml.safe_load in release jobs
OpenDev bumped the default nodeset to ubuntu noble. This change has
caused PyYaml to fail with this error with using yaml.load(input):

    File "/home/zuul/scripts/release-tools/process_release_requests.py", line 171, in process_release_requests
      deliverable_data = yaml.load(f.read())
                         ^^^^^^^^^^^^^^^^^^^
  TypeError: load() missing 1 required positional argument: 'Loader'

The issue is that yaml.load is not safe and they deprecated the naive
use of yaml.load(). Instead you have to specify a loader and opt into
unsafe behavior (or opt into safe behavior). We can avoid this entirely
by using yaml.safe_load() instead so we make that change here.

Change-Id: I3bbca9af450cf2b0422ccc568a49b2b672400a59