Omnibus fixes for running jobs on Noble with Python3.12
We bumped the default nodeset to ubuntu-noble which bumped our python
version up to python3.12. We were previously pinning hacking to a
version that pulls in a version of flake8 that isn't compatible with
python3.12 due to entrypoint importlib_metadata compatibility issues.

We fix that by updating hacking to lastest, but that introduces a new
rule that requires we have whitespace after a return statement. We also
fix that in the one location that was a problem.

Then separately ssl.wrap_context has been removed in python 3.12 so we
switch that out with SSLContext.wrap_context instead when checking irc
access in the irc access check job.

Change-Id: Id4e3c6f35ffcf5cea302aa1bd8b126affa4eeb07