Run openstack-zuul-jobs-linters on Jammy
There are problems with running this job on Noble due to Python 3.12
incompatibities with our ancient ansible-lint setup. Pin the job back
for now so taht we can land things after the Noble default nodeset
change.

Followups will deal with the ansible-lint setup to make it more
resilient to newer python (or possibly remove it or reduce its scope).

Change-Id: I0bc15907bb672934942f368a4c0d868d3f6fbee4