Try booting nodes in Rackspace Flex
Our expected Nodepool images are present in the raxflex-sjc3
provider now, so raise max-servers to 1 in order to get some sample
builds we can analyze for obvious issues like connectivity.

Depend on the subnet resizing change since that might otherwise be
disruptive if it deployed while jobs are running there.

Depends-On: https://review.opendev.org/927813
Change-Id: If8b39b53608188e5881a273e6d092981a5871e84