Increase raxflex-sjc3 max-servers to 32
Looking at `openstack limits show --absolute` we seem to currently
have sufficient memory quota to take this up to 32x 8GB RAM
instances (quotas for CPU and total instances are higher so RAM is
the limiting factor at the moment).

Things seem to be going fine at 20 nodes for the past few hours, so
let's go ahead and dial it up to maximum for the weekend and see how
it works out before we let our Rackspace contacts know we're ready
for additional quota.

Change-Id: I1d61da4ca2a1a5963e0cdcff99fdf4f41aea5920