Use ubuntu-jammy for propose-update-constraints
Recently propose-update-constraints was set to use ubuntu-noble, which
works fine for most of the cases, especially with current master branch
(2024.2 Dalmatian), as folks are preparing their projects to be able to
run on Ubuntu 24.04 Noble Numbat (which will be available in 2025.1
Epoxy [1]). The problem is that stable branches can still contain old
things, and for example outdated bindep.txt's cause the job to fail on
ubuntu-noble [2][3]. To avoid this, let's pin the job to ubuntu-jammy
instead.

[1] https://governance.openstack.org/tc/reference/runtimes/2025.1.html
[2] https://lists.openstack.org/archives/list/release-job-failures@lists.openstack.org/thread/WGJQE34T33MVNYX3MXLIKTC56E4F3NX7/
[3] https://zuul.opendev.org/t/openstack/build/1952087280a946b5b04d5854c06da6ae

Change-Id: Id1bbfa060664d4689ba696a511fdef7b24c9fbfc