Convert python2 dib element scripts to python3
As part of the move to zuul jobs building images we've discovered we
have a couple of scripts that are invoked by python2. Python2 isn't
available be default in the test images and python2 is old enough now
that we should just use python3. This change updates things to make it
so.

We also drop initialize-urandom from our arm builds. We haven't used
this with x86 in some time and it was only really an issue for xenial
anyway. If x86 xenial doesn't need it then arm shouldn't either.

Change-Id: I3821336758b13051e9b331b8feb73fb1f8f1184a