Set launch-timeout on nodepool providers
The nodepool openstack provider implementation has a launch timeout
default of 3600 seconds or one hour. This is problematic for us because
a node request will be attempted three times by a provider before being
passed onto the next provider. This means we may wait up to three hours
per failed provider to launch a single node.

Fix this by setting a timeout of 10 minutes on the three providers that
didn't already set a lower timeout (raxflex and the two vexxhost
providers). Looking at grafana graphs for time to ready this should be
plenty of time for normal booting conditions and reduces the time we
wait from 3 hours per provider to 30 minutes.

Change-Id: I0a4af9f5519ff2b64c737b1822590fbb2608e8bb