Set bindep profiles for openstack release doc publication
The openstack-tox-docs job which build the release docs in check and
gate set the `compile doc` bindep profiles. This was used to
successfully install graphviz and generate a diagram with this tool in
check and gate jobs for openstack/releases. Then when we got to building
the published artifacts we installed the default `compile test` profile
which lacked graphviz and produced a string output of the source code
for our diagram. Not what we wanted.

Update the job to use a matching bindep profile so that we get
consistent results between check, gate, and release-post. A change to
openstack/releases will need to land after this update to project-config
in order to rebuild things there.

Change-Id: I36335bfb70c8b6e6f1a84dbb59498b8c2da05825