Switch the remaining opendev zuul tenants to ansible 9 by default
We switched zuul and opendev a little while back and fixed the bugs that
we discovered. Since then OpenStack has produced its 2024.2 Dalmatian
release which makes this a good time to update everything else.

I did check that devstack+tempest seem to generally work in this change
[0] as they do a bunch of non trivial stuff that should help ensure
general compatibility. The change got a -1 but this was due to a known
bug with nova jobs. It seems to work generally and did not hit ansible
update problems.

[0] https://review.opendev.org/c/openstack/tempest/+/929684

Change-Id: I953590525f1ca3968ebd16839d43d91741091dd7