Add the vexxhost/project-config repo
We want to make some changes to opendev/project-config which we do
not yet want to apply to all tenants, so fork a copy of it for
the vexxhost tenant which is the only tenant other than opendev
that uses it.

Temporarily disable the zuul config check since that is expected in
this case.

Change-Id: Iec16d0a92bff3a33751548d717ba93d1be3eeeb8