Set raxflex max-servers to 0
Rax flex's mirror node lost its apache and openafs cache cinder volume.
We're getting errors like:

  I/O error, dev vdd, sector 0 op 0x0:(READ) flags 0x0 phys_seg 1 prio class 0

And it doesn't even recognize that there is an lvm partition scheme
present on the device. This is preventing the mirror from functioning.

Drop max servers to 0 as we need a working mirror for jobs to run
successfully.

Change-Id: Ica1970c2ac117104ea5a34e23e82ad30376588c7