Add repository for OSA's httpd role
In order to unify approach on handling Apache Web server deployment
and configuration team come up with idea to produce a separate role
for that as we have failed with finding existing one which would meet
our intended "pluggable" design.

Based on that a new repo/project is needed as we intend to make
resulting role not opinionated and re-usable outside of OSA scope.

Needed-By: https://review.opendev.org/c/openstack/governance/+/935694
Change-Id: Iae39bdd3084c5a74a8e2cc0d2b3061cc51571116