update watcher repos to use -release group for stable branch
Watcher has had watcher-release and watcher-dashboard-release
for a long time but they have not been correctly configured
in the gerrit ACLs.

This change updates the relevent acl config to allow
watcher-release member to merge patches on stable branches for
watcher and python-watcherclint, and watcher-dashboard-release
to manage the watcher-dashboard stable branches.

The stable-maint-core group is a member of both -release groups
and will inherit review rights as a result of this change.

The watcher tempest and spec configs are not updated as they
are not branched.

Change-Id: I3f5e4b9d148ec01a6603458a8e9abba2c2e57152