Use the new infra-prod locking and dependency setup
We trigger some infra-prod deploy jobs from changes to project-config.
We recently updated system-config to run jobs with new locking and
depdendency setups and project-config needs to be updated to ensure
everything works as expected on this side of the house.

In particular normal infra-prod jobs no longer update repos on bridge.
Instead infra-prod-bootstrap-bridge does this then pauses while holding
a lock then pauses while the other infra-prod jobs in the buildset run.
This ensures the jobs belonging to that buildset run with the correct
git repo state and run with exclusivity compared to all other infra-prod
deploy/periodic/opendev-prod-hourly buildsets.

Change-Id: I64fc0fcc15e23bdc9b04eb23a75d77d576dc23bf